requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
python-decouple
psycopg2-binary
gunicorn
mailjet-rest
Babel
reportlab
openpyxl
pypandoc
Pillow
django-tenants
django_apscheduler
APScheduler 
tzlocal
djangorestframework-simplejwt
pandas
